refactor: 불필요한 import 삭제, 디렉토리명 변경, 쿼리 타임 수정 (#141)
* delete: 사용하지 않는 import 삭제

* refactor: 쿼리 타임 수정

* refactor: 스케줄러 디렉토리 명 변경
