+++ src/main/java/kr/allcll/backend/domain/timetable/schedule/dto/ScheduleCreateRequest.java
@@ -1,8 +1,6 @@
 package kr.allcll.backend.domain.timetable.schedule.dto;
 
 import java.util.List;
-import kr.allcll.backend.domain.timetable.schedule.ScheduleType;
-import org.antlr.v4.runtime.misc.NotNull;
 
 public record ScheduleCreateRequest(
     String scheduleType,
